requirements.txt
langfuse
python-dotenv
google-genai
pytest
